requirements.txt
anyio==3.6.1
asgiref==3.5.2; python_full_version >= "3.6.1" and python_version >= "3.7"
certifi==2022.5.18.1; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.1"
charset-normalizer==2.0.12; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.1"
click==8.1.3; python_full_version >= "3.6.1" and python_version >= "3.7"
colorama==0.4.4
dnspython==2.2.1; python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.6.1"
email-validator==1.2.1; python_full_version >= "3.6.1"
fastapi==0.78.0; python_full_version >= "3.6.1"
h11==0.13.0; python_full_version >= "3.6.1" and python_version >= "3.7"
httptools==0.4.0
idna==3.3; python_full_version >= "3.6.2" and python_version >= "3.7" and python_version < "4"
itsdangerous==2.1.2; python_full_version >= "3.6.1" and python_version >= "3.7"
jinja2==3.1.2; python_full_version >= "3.6.1" and python_version >= "3.7"
markupsafe==2.1.1; python_full_version >= "3.6.1" and python_version >= "3.7"
orjson==3.7.2; python_full_version >= "3.6.1" and python_version >= "3.7"
pydantic==1.9.1; python_full_version >= "3.6.1"
python-dotenv==0.20.0
python-multipart==0.0.5; python_full_version >= "3.6.1"
pyyaml==6.0
requests==2.28.0; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.1"
six==1.16.0; python_full_version >= "3.6.1"
sniffio==1.2.0; python_version >= "3.6" and python_full_version >= "3.6.2"
starlette==0.19.1; python_version >= "3.6" and python_full_version >= "3.6.1"
typing-extensions==4.2.0; python_version >= "3.7" and python_full_version >= "3.6.1"
ujson==5.3.0; python_full_version >= "3.6.1" and python_version >= "3.7"
urllib3==1.26.9; python_full_version >= "3.6.1" and python_version < "4" and python_version >= "3.7"
uvicorn==0.17.6; python_full_version >= "3.6.1" and python_version >= "3.7"
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
